deploy/requirements.txt
psycopg2-binary==2.9.5
PyYAML==6.0.1
seaborn==0.13.2
python-dateutil==2.8.2
matplotlib==3.8.2
ruamel.yaml==0.18.5
